@@ -1,18 +1,33 @@
-[package]
-name = "ccsync"
+[workspace]
+members = [
+    "crates/ccsync",
+    "crates/ccsync-cli",
+]
+resolver = "2"
+
+[workspace.package]
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 authors = ["ccsync contributors"]
-description = "Claude Configuration Synchronization Tool - Sync agents, skills, and commands between global and project directories"
 license = "MIT OR Apache-2.0"
-repository = "https://github.com/yourusername/ccsync"
-keywords = ["cli", "sync", "configuration", "claude"]
-categories = ["command-line-utilities", "development-tools"]
+repository = "https://github.com/onsails/ccsync"
 
-[dependencies]
-clap = { version = "4.5", features = ["derive", "cargo"] }
-anyhow = "1.0"
+[workspace.dependencies]
+# Core dependencies with tilde requirements for patch compatibility
+clap = { version = "~4.5.28", features = ["derive", "cargo"] }
+anyhow = "~1.0.95"
 
-[dev-dependencies]
-assert_cmd = "2.0"
-predicates = "3.1"
+# Dev dependencies
+assert_cmd = "~2.0.17"
+predicates = "~3.1.3"
+
+[workspace.lints.rust]
+unsafe_code = "forbid"
+unused_imports = "warn"
+unused_mut = "warn"
+dead_code = "warn"
+
+[workspace.lints.clippy]
+all = "warn"
+pedantic = "warn"
+nursery = "warn"
